octofit-tracker/backend/requirements.txt
django>=4.1.13
djangorestframework>=3.14.0
djongo==1.3.6
pymongo==3.12.3
django-cors-headers>=4.3.1
